requirements.txt
Django==3.2.9
djangorestframework==3.12.4
Scrapy==2.5.1
crochet
scrapyd~=1.2.1
python-scrapyd-api~=2.1.2
django-debug-toolbar~=3.2.2